OTG: move UsbConnectionDispatcher to utls package
