--- OpenKeychain/src/main/java/org/sufficientlysecure/keychain/smartcard/OnDiscoveredUsbDeviceListener.java
@@ -1,7 +1,0 @@
-package org.sufficientlysecure.keychain.smartcard;
-
-import android.hardware.usb.UsbDevice;
-
-public interface OnDiscoveredUsbDeviceListener {
-    void usbDeviceDiscovered(UsbDevice usbDevice);
-}
